:boom: Refactor Client class to take api_root/base_url as only required arg
Client initializing is being made much more straightforward, without a config registry
or aliases of sorts.

This also removes the `Client.from_url` functionality, as it was always founded
on assumptions about the endpoint structure and usage of UUIDs, which is definitely
not always the case (see: Haal Centraal BRP bevragen).

Users of this functionality now need to keep track of the APIs they're
consuming (which is required for auth anyway) and should figure out which API root
is relevant for a given resource URL.